Delete a bunch of unused configs
Bit of a change in philosophy here. I'd previously thought to keep
these older configuration files around, just in case, but it's not
likely that I turn back to Linux and XMonad in the near future, so let
me delete them and keep the directory tree clean. They're always
available in the `git` history, after all.

 #alacritty #dmenu #dvtm #emacs #i3 #osx-terminal #wallpapers
 #xfce4-terminal #xmobar #xmonad